solve some security review and bad smells

+++ src/main/java/application/services/DatabaseService.java
@@ -4,7 +4,10 @@ import application.database.QueryResultHandler;
 import java.io.Serializable;
 import java.lang.reflect.Field;
 import java.sql.*;
+import java.util.Arrays;
+import java.util.List;
 import java.util.Objects;
+import java.lang.reflect.AccessibleObject;
 import static application.services.MessagesGenerator.LOGGER;
 
 
@@ -13,6 +16,7 @@ public class DatabaseService implements Serializable {
   private static final String LITERAL_FOR_PROFILE = "profile";
   private static final String DATABASE_USER = System.getenv("DATABASE_USER");
   private static final String DATABASE_PASSWORD = System.getenv("DATABASE_PASSWORD");
+  private final String LITERAL_FOR_REQUEST_TABLE = "Request";
   private static Connection connection;
 
   static {
@@ -36,33 +40,36 @@ public class DatabaseService implements Serializable {
 
   public <T> T executeQuery(String query, QueryResultHandler<T> resultHandler) throws SQLException{
 
-     ResultSet resultSet;
-     PreparedStatement statement ;
+    ResultSet resultSet;
+    PreparedStatement statement ;
 
-     statement = connection.prepareStatement(query);
-     resultSet = statement.executeQuery();
-     return resultHandler.handle(resultSet);
+    statement = connection.prepareStatement(query);
+    resultSet = statement.executeQuery();
+    return resultHandler.handle(resultSet);
 
 
   }
-
-
-  public  <T> boolean addObject(T object, String tableName) throws SQLException{
-
+  public <T> boolean addObject(T object, String tableName) throws SQLException {
     StringBuilder insertQuery = new StringBuilder("INSERT INTO " + tableName + " (");
 
     Field[] fields = object.getClass().getDeclaredFields();
     int fieldsLength = (Objects.equals(tableName, "Request")) ? fields.length - 1 : fields.length;
+
+    boolean[] fieldAccessible = new boolean[fieldsLength];
+
     for (int i = 0; i < fieldsLength; i++) {
-      if(fields[i].getName().equals(LITERAL_FOR_PROFILE))
+      if (fields[i].getName().equals(LITERAL_FOR_PROFILE)) {
         insertQuery.append("profileId");
-      else
+      } else {
         insertQuery.append(fields[i].getName());
+      }
+
+      fieldAccessible[i] = fields[i].canAccess(object);
+
       if (i < fieldsLength - 1) {
         insertQuery.append(", ");
       }
     }
-
 
     insertQuery.append(") VALUES (");
     for (int i = 0; i < fieldsLength; i++) {
@@ -73,36 +80,52 @@ public class DatabaseService implements Serializable {
     }
     insertQuery.append(")");
 
-
-    try{
+    try {
       PreparedStatement statement = connection.prepareStatement(insertQuery.toString());
       for (int i = 0; i < fieldsLength; i++) {
-        fields[i].setAccessible(true);
-            statement.setObject(i + 1, fields[i].get(object).toString());
+        if (!fieldAccessible[i]) {
+          fields[i].setAccessible(true);
+        }
+        statement.setObject(i + 1, fields[i].get(object).toString());
       }
 
       statement.executeUpdate();
       statement.close();
       return true;
 
-    }catch ( Exception e ){
+    } catch (Exception e) {
       return false;
+    } finally {
+      for (int i = 0; i < fieldsLength; i++) {
+        if (!fieldAccessible[i]) {
+          fields[i].setAccessible(false);
+        }
+      }
     }
-
   }
 
 
-  public boolean deleteObject(int id, String tableName) throws SQLException{
+
+  public boolean deleteObject(int id, String tableName) throws SQLException {
+    if (!isValidTableName(tableName)) {
+      throw new IllegalArgumentException("Invalid table name");
+    }
 
     String deleteQuery = "DELETE FROM " + tableName + " WHERE id = ?";
-    try ( PreparedStatement statement = connection.prepareStatement(deleteQuery) ) {
+    try (PreparedStatement statement = connection.prepareStatement(deleteQuery)) {
       statement.setInt(1, id);
       statement.executeUpdate();
       return true;
     }
   }
 
-  public <T> void updateObject(T object, String tableName, String primaryKeyField) throws Exception {
+  private boolean isValidTableName(String tableName) {
+    List<String> allowedTableNames = Arrays.asList("Product", "Profile", "Request", "user");
+    return allowedTableNames.contains(tableName);
+  }
+
+
+  public <T> void updateObject(T object, String tableName, String primaryKeyField) throws SQLException, NoSuchFieldException, IllegalAccessException {
     String updateQuery = buildUpdateQuery(object, tableName, primaryKeyField);
     try (PreparedStatement statement = connection.prepareStatement(updateQuery)) {
       setParameters(statement, object, primaryKeyField);
@@ -114,7 +137,7 @@ public class DatabaseService implements Serializable {
     StringBuilder updateQuery = new StringBuilder("UPDATE " + tableName + " SET ");
 
     Field[] fields = object.getClass().getDeclaredFields();
-    int fieldsLength = (Objects.equals(tableName, "Request")) ? fields.length - 1 : fields.length;
+    int fieldsLength = (Objects.equals(tableName, LITERAL_FOR_REQUEST_TABLE)) ? fields.length - 1 : fields.length;
 
     for (int i = 0; i < fieldsLength; i++) {
       if (!fields[i].getName().equals(primaryKeyField)) {
@@ -138,7 +161,7 @@ public class DatabaseService implements Serializable {
 
   private <T> void setParameters(PreparedStatement statement, T object, String primaryKeyField) throws IllegalAccessException, NoSuchFieldException, SQLException {
     Field[] fields = object.getClass().getDeclaredFields();
-    int fieldsLength = (Objects.equals("Request", primaryKeyField)) ? fields.length - 1 : fields.length;
+    int fieldsLength = (Objects.equals(LITERAL_FOR_REQUEST_TABLE, primaryKeyField)) ? fields.length - 1 : fields.length;
 
     int paramIndex = 1;
     for (int i = 0; i < fieldsLength; i++) {
@@ -149,18 +172,33 @@ public class DatabaseService implements Serializable {
     }
 
     Field primaryKey = object.getClass().getDeclaredField(primaryKeyField);
-    primaryKey.setAccessible(true);
-    statement.setObject(paramIndex, primaryKey.get(object));
-  }
 
-  private void setParameter(PreparedStatement statement, Field field, Object object, int paramIndex) throws IllegalAccessException, SQLException {
-    field.setAccessible(true);
-    if (field.getName().equals(LITERAL_FOR_PROFILE) || field.getName().equals("role")) {
-      statement.setObject(paramIndex, field.get(object).toString());
-    } else {
-      statement.setObject(paramIndex, field.get(object));
+    AccessibleObject.setAccessible(new AccessibleObject[]{primaryKey}, true); // Set accessibility to true
+
+    try {
+      statement.setObject(paramIndex, primaryKey.get(object));
+    } finally {
+      AccessibleObject.setAccessible(new AccessibleObject[]{primaryKey}, false); // Set accessibility back to false
     }
   }
+
+
+
+  private void setParameter(PreparedStatement statement, Field field, Object object, int paramIndex) throws IllegalAccessException, SQLException {
+    AccessibleObject.setAccessible(new AccessibleObject[]{field}, true); // Set accessibility to true
+
+    try {
+      if (field.getName().equals(LITERAL_FOR_PROFILE) || field.getName().equals("role")) {
+        statement.setObject(paramIndex, field.get(object).toString());
+      } else {
+        statement.setObject(paramIndex, field.get(object));
+      }
+    } finally {
+      AccessibleObject.setAccessible(new AccessibleObject[]{field}, false); // Set accessibility back to false
+    }
+  }
+
+
 
 
 }
